Restructure project folder structure to allow for unit tests in future update.
