+++ dev_main.py
@@ -1,0 +1,4 @@
+from article_overload import basic_bot
+
+if __name__ == "__main__":
+    basic_bot.main()
